refactor(platform): promote DNS / Interface / NetworkStats / RouteTable ports (plan 003 U2)
Adds the remaining four capability ports in vortix-core and relocates
every platform-specific impl into the dedicated platform crates. Cycle
removal lands in U4; this commit removes all binary-side OS-impl code.

New port traits in vortix-core/src/ports/:
- dns.rs — DnsResolver
- interface.rs — Interface (renamed from InterfaceDetector)
- network_stats.rs — NetworkStats (renamed from NetworkStatsProvider)
- route_table.rs — RouteTable (new; default-gateway lookup that today
  lives free-function-style in core/network_monitor.rs)

Impl relocation:
- vortix-platform-linux/src/{dns,interface,network_stats,route_table}.rs
  — moved from crates/vortix/src/platform/linux/*. Embeds the constants
  the impls need (RESOLV_CONF_PATH, PROC_NET_DEV_PATH) locally.
- vortix-platform-macos/src/{dns,interface,network_stats,route_table}.rs
  — moved from crates/vortix/src/platform/macos/*. Embeds
  WIREGUARD_RUN_DIR and the netstat timeout locally.
- The /proc/net/dev and `ip addr` parsers that lived in
  core/telemetry.rs migrate alongside the network_stats / interface impls
  (with their tests).

Binary-side shims:
- crates/vortix/src/platform/mod.rs — aliases the new trait names back
  to the legacy ones (Firewall, DnsResolver, InterfaceDetector,
  NetworkStatsProvider) so existing call sites keep working until U7.
- crates/vortix/src/platform/{linux,macos}/mod.rs — re-export the
  vortix-platform-* submodules so `crate::platform::linux::dns::LinuxDns`
  etc. continue to resolve.
- crates/vortix/src/platform/{linux,macos}/{dns,interface,network}.rs
  deleted (now sourced from the platform crates).

Verification:
- cargo build --workspace --all-targets ✓
- cargo test --workspace ✓
- cargo clippy --all-targets -- -D warnings ✓
- cargo xtask check-subprocess ✓
