+++ crates/vortix-platform-linux/src/dns.rs
@@ -1,8 +1,9 @@
 //! Linux DNS resolver using resolvectl, nmcli, and /etc/resolv.conf.
 
-use crate::constants;
-use crate::platform::DnsResolver;
+use vortix_core::ports::dns::DnsResolver;
 use vortix_process::CommandSpec;
+
+const RESOLV_CONF_PATH: &str = "/etc/resolv.conf";
 
 /// Linux DNS resolution with fallback chain:
 /// 1. `resolvectl` (systemd-resolved)
@@ -76,7 +77,7 @@ fn try_get_dns_nmcli() -> Option<String> {
 
 /// Try to get DNS from /etc/resolv.conf (universal fallback).
 fn try_get_dns_resolv_conf() -> Option<String> {
-    let content = std::fs::read_to_string(constants::RESOLV_CONF_PATH).ok()?;
+    let content = std::fs::read_to_string(RESOLV_CONF_PATH).ok()?;
     for line in content.lines() {
         let trimmed = line.trim();
         if trimmed.starts_with("nameserver") {
