+++ crates/vortix/src/platform/linux/mod.rs
@@ -1,11 +1,7 @@
-//! Linux platform implementations.
+//! Linux platform implementations — thin re-exports.
 //!
-//! Uses iptables/nftables, /proc/net/dev, ip addr, and resolvectl.
+//! The actual impl code lives in `vortix-platform-linux` per plan 003 U1/U2.
+//! Submodule aliases here keep existing `crate::platform::linux::*` paths
+//! resolving until plan 003 U7 swaps consumers over to `&Platform`.
 
-pub mod dns;
-pub mod interface;
-pub mod network;
-
-// Killswitch firewall impl now lives in `vortix-platform-linux::firewall`
-// (plan 003 U1).
-pub use vortix_platform_linux::firewall;
+pub use vortix_platform_linux::{dns, firewall, interface, network_stats as network, route_table};
